[REFACTOR][I] Split off document handling from ProjectAPI
Splits of the methods exclusively dealing with documents from the class
ProjectAPI and moves them to the newly created class DocumentAPI.
This is done as there are some components that are needed in the plugin
context that rely on some of the document method. By splitting them of,
this allows us to move the project dependent ProjectAPI to the session
context.

Subsequently makes the moved methods static as they are not dependent
on any other non-static resource.
