+++ intellij/src/saros/intellij/editor/ProjectAPI.java
@@ -1,6 +1,5 @@
 package saros.intellij.editor;
 
-import com.intellij.openapi.application.ModalityState;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
@@ -9,11 +8,9 @@ import com.intellij.openapi.fileEditor.OpenFileDescriptor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Computable;
 import com.intellij.openapi.vfs.VirtualFile;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import saros.intellij.filesystem.Filesystem;
 
-/** IntellIJ API for project-level operations on editors and documents. */
+/** IntellIJ API for project-level operations on editors. */
 public class ProjectAPI {
   private FileDocumentManager fileDocumentManager;
 
@@ -69,29 +66,6 @@ public class ProjectAPI {
   }
 
   /**
-   * Returns a document for the given file.
-   *
-   * @param file the <code>VirtualFile</code> for which the document is requested
-   * @return a <code>Document</code> for the given file or <code>null</code> if the file does not
-   *     exist, could not be read, is a directory, or is to large
-   */
-  @Nullable
-  public Document getDocument(@NotNull final VirtualFile file) {
-    if (!file.exists()) {
-      return null;
-    }
-
-    return Filesystem.runReadAction(
-        new Computable<Document>() {
-
-          @Override
-          public Document compute() {
-            return fileDocumentManager.getDocument(file);
-          }
-        });
-  }
-
-  /**
    * Closes the editor for the given file in the UI thread.
    *
    * @param file
@@ -135,51 +109,5 @@ public class ProjectAPI {
    */
   public VirtualFile[] getSelectedFiles() {
     return editorFileManager.getSelectedFiles();
-  }
-
-  /**
-   * Saves the given document in the UI thread.
-   *
-   * @param doc
-   */
-  public void saveDocument(final Document doc) {
-    Filesystem.runWriteAction(
-        new Runnable() {
-
-          @Override
-          public void run() {
-            fileDocumentManager.saveDocument(doc);
-          }
-        },
-        ModalityState.NON_MODAL);
-  }
-
-  /**
-   * Reloads the current document in the UI thread.
-   *
-   * @param doc
-   */
-  public void reloadFromDisk(final Document doc) {
-    Filesystem.runReadAction(
-        new Runnable() {
-
-          @Override
-          public void run() {
-            fileDocumentManager.reloadFromDisk(doc);
-          }
-        });
-  }
-
-  /** Saves all documents in the UI thread. */
-  public void saveAllDocuments() {
-    Filesystem.runWriteAction(
-        new Runnable() {
-
-          @Override
-          public void run() {
-            fileDocumentManager.saveAllDocuments();
-          }
-        },
-        ModalityState.NON_MODAL);
   }
 }
